requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.2
transformers>=4.21.0
PyMuPDF>=1.19.6
numpy>=1.21.0
torch>=1.10.0